backend/requirements.txt
fastapi
uvicorn
pandas
python-multipart
psycopg2-binary
sqlalchemy
openai
python-dotenv
google-genai
pytest
cryptography
